refactor: NFTList into AssetList
