+++ webapp/src/components/AssetList/index.ts
@@ -1,0 +1,2 @@
+import AssetList from './AssetList.container'
+export { AssetList }
