--- webapp/src/components/NFTList/index.ts
@@ -1,2 +1,0 @@
-import NFTList from './NFTList.container'
-export { NFTList }
